- various changes to reduce warnings, most being addition of const modifiers, and use defined default values - fixed last contact behavior in lua code - added missing files in svn

+++ bot/misc.c
@@ -166,7 +166,7 @@ char	*time2small(long time)
 
 char	*idle2str(long time)
 {
-	int	days,
+	unsigned short	days,
 		hours,
 		mins,
 		secs;
